rhogen: split into two parts, move module logic to separate file
